# openapi.yaml
openapi: 3.0.0
info:
  title: Feedback 360 Survey API
  version: 2.0.0
  description: |
    This API specification contains information around the Feedback 360 Survey
    API developed by the App Dev Practice.

    The Feedback 360 Survey API enables the management of feedback surveys for
    customer engagements. A feedback survey is N/A to 5 point scale of competency around
    the technologies used for a project. Each added employee will be able to enhance the
    survey by adding technologies, softskills, and freetext for each of his or her
    colleagues. The will enable Red Hat to better understand multiple aspects of
    a consultant's skillset.
  contact:
    email: appdevpractice@redhat.com
  license:
    name: Apache 2.0
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
...
servers:
  - url: "http://localhost:3007"
    description: LOCAL
  - url: "http://survey-service-consultant-360-dev.apps.shared-dev.dev.openshift.opentlc.com"
    description: Spring Boot Service that implements the Feedback 360 Survey API on DEV Cluster
  - url: "http://survey-service-consultant-360-prod.apps.shared-dev.dev.openshift.opentlc.com"
    description: Spring Boot Service that implements the Feedback 360 Survey API on PROD Cluster
paths:
...
  /skills:
    get:
      tags:
        - skills
      parameters:
        - name: filterBy
          description: Field by which to filter
          in: query
          schema:
            type: string
          example: skill
        - name: filterString
          description: String to filter on, query string
          in: query
          schema:
            type: string
          example: Openshift
        - name: sortBy
          description: Field by which to sort
          in: query
          schema:
            type: string
        - name: sortOrder
          description: Sort Order
          in: query
          schema:
            type: string
            enum:
              - ASC
              - DESC
        - name: offset
          description: Page offset
          schema:
            type: integer
            format: int32
          in: query
        - name: maxResults
          description: Maximum number of results to return, defaults to 20
          schema:
            type: integer
            format: int32
          example: 20
          in: query
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Skill"
          description: List of Skills that are available.
      operationId: getAllSkills
      summary: Fetch all the skills available to be added on a survey submission.
      description: Returns all skills that are active in the Skills table.
...
  "/rmo/profile/{employeeId}/skills":
    get:
      tags:
        - RMOUserProfile
      parameters:
        - name: employeeId
          description: Kerberos ID of the employee.
          schema:
            type: string
            example: ajones
          in: path
          required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array 
                items:
                  $ref: "#/components/schemas/EmployeeSkillHistory"
          description: A list of all the project and corresponding skills associated with an employee.
        "404":
          description: No projects were found for the given employee.
      operationId: getEmployeeSkillsHistory
      summary: Find a list of all the project and corresponding skills associated with an employee.
      description: Return a Map<project,skills> of all the projects an employee is part of and a list of skills that the employee used at each project.
  "/rmo/profile/{employeeId}/aggregateRatings":
    get:
      tags:
        - RMOUserProfile
      parameters:
        - name: employeeId
          description: Kerberos ID of the employee.
          schema:
            type: string
            example: ajones
          in: path
          required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmployeeAggregateSkillRatings"
          description: Aggregate Skill Ratings for each skill that the employee has be rated on 3 or more times.
        "404":
          description: No skills with more than three ratings were found for the given employee.
      operationId: getEmployeeAggregateSkillRatings
      summary: Find a list ggregate Skill Ratings for each skill that the employee has be rated on 3 or more times.
      description: Return a Map<skill,rating> of all the the skills that the employee has been rated on 3 or more times.
components:
  schemas:
...
    Skill:
      type: object
      x-java-class-annotations:
        - "@javax.persistence.Entity"
        - '@javax.persistence.Table(name = "skills")'
      properties:
        id:
          description: A GUID that uniquely identifies a skill.
          type: string
          format: uuid
          example: d567521f-13a6-4237-936c-40bbbf388bc5
          x-java-field-annotations:
            - "@javax.persistence.Id"
            - '@javax.persistence.GeneratedValue(generator = "UUID")'
            - '@org.hibernate.annotations.GenericGenerator(name = "UUID",strategy = "org.hibernate.id.UUIDGenerator")'
            - '@javax.persistence.Column(name = "id", updatable = false, nullable = false)'
        skillsBaseId:
          description: An id from skills base
          type: integer
          example: 1234
        skill:
          description: The name of the skill.
          type: string
          example: Crucial Conversations
        description:
          description: A description of the skill.
          type: string
          example: The ability to have crucial conversations with clients.
          x-java-field-annotations:
            - "@javax.persistence.Column(length=512)"
        category:
          description: The general category of the skill.
          type: string
          example: leadership
        active:
          description: Boolean to state if this skill is active
          type: boolean
          example: true
          default: true
...
    EmployeeSkillHistory:
      type: object
      properties:
        project:
          description: Name of project that consultant was part of
          type: string
          example: "Bank of America"
        skills:  
          description:  Array of skills that the consultant used at the project
          type: array
          items:
            type: string
            example: "Spring"
    EmployeeAggregateSkillRatings:
      type:  object
      properties:
        skill:
          description:  Skill that the consultant was rated on
          type: string
          example: "Spring"
        rating:
          description:  Aggregate skill rating for the above skill 
          type: integer
          example: 5
...